Refactor admin and tenders test files; add new tests for pagination, filtering, and analytics functionality
