--- cypress/e2e/tests/tenders/c260_viewTendersInList.cy.ts
@@ -1,11 +1,0 @@
-import TendersMapPage from "../../pages/tendersPages/TendersMapPage";
-import Header from "../../components/Header";
-
-describe("Tenders functionality", () => {
-  it("C260 View the tenders in the list", () => {
-    TendersMapPage.visit();
-    Header.clickTendersBtn();
-    TendersMapPage.clickShowListBtn();
-    TendersMapPage.verifyFirstAndLastTenderCardElements();
-  });
-});
